Changed int to int32_t and long long to uint64_t
Bug fixed : scanf_s(_ ,_ ,MAXINPUT) is replaced by   scanf_s(_ ,_ ,MAXINPUT+1) in Calculator.c

+++ calculator.h
@@ -14,6 +14,7 @@
 #include <limits.h> //for INT_MAX
 #include <string.h> //for strspn and strlen
 #include <math.h> // for pow
+#include <stdint.h>
 #define MAXSIZEINPUT 10 //because the INT_MAX is 10 digit, if we have more that 10 digit, we have an overflow for sure.
 
 enum errorCodes {
@@ -21,23 +22,23 @@ enum errorCodes {
 };
 
 char* getErroryName(enum Days day);
-int add(int firstInt, int secondInt);
-double calculate(char* firstStr, char operator, char* secondStr, int* firstNbrInt, int* secondNbrInt);
-int checkOverflowAndReturnResult(long long result);
-double divide(int firstInt, int secondInt);
-int euclideanDivision(int first, int second);
-double executeOperation(int first, char operator, int second);
-int inputIsValid(char* firstStr, char* secondStr, int* firstNbrInt, int* secondNbrInt);
-int modulo(int firstInt, int secondInt);
-int multiply(int firstInt, int secondInt);
-double nthRoot(int firstInt, int secondInt);
-int power(int firstInt, int secondInt);
-int mainForDemo();
+int32_t add(int32_t firstInt, int32_t secondInt);
+double calculate(char* firstStr, char operator, char* secondStr, int32_t* firstNbrInt, int32_t* secondNbrInt);
+int32_t checkOverflowAndReturnResult(uint64_t result);
+double divide(int32_t firstInt, int32_t secondInt);
+int32_t euclideanDivision(int32_t first, int32_t second);
+double executeOperation(int32_t first, char operator, int32_t second);
+int32_t inputIsValid(char* firstStr, char* secondStr, int32_t* firstNbrInt, int32_t* secondNbrInt);
+int32_t modulo(int32_t firstInt, int32_t secondInt);
+int32_t multiply(int32_t firstInt, int32_t secondInt);
+double nthRoot(int32_t firstInt, int32_t secondInt);
+int32_t power(int32_t firstInt, int32_t secondInt);
+int32_t mainForDemo();
 void printResult(double result);
-long long stringConvertToLong(char* stringNbr);
-int substract(int firstInt, int secondInt);
-int verifyNumbersAreInt(long long firstLongNbr, long long secondLongNbr);
+uint64_t stringConvertToLong(char* stringNbr);
+int32_t substract(int32_t firstInt, int32_t secondInt);
+int32_t verifyNumbersAreInt(uint64_t firstLongNbr, uint64_t secondLongNbr);
 void printError();
-int mainForDemo();
+int32_t mainForDemo();
 
 #endif /* calculator_h */
